+++ app/containers/Main/routes.js
@@ -13,8 +13,8 @@ import IssuesDetail from 'containers/Issues/Detail';
 import Jobs from 'containers/Jobs';
 import Main from 'containers/HomePage/Loadable';
 import Overview from 'containers/Overview';
-import ReposAdd from 'containers/Organizations/Add';
-import ReposDetail from 'containers/Organizations/Detail';
+import ReposAdd from 'containers/Repos/Add';
+import ReposDetail from 'containers/Repos/Detail';
 import Settings from 'containers/Settings';
 import SignIn from 'containers/Signin';
 import Stats from 'containers/Stats/Loadable';
